Don't destroy the path if VS is not found

+++ scripts/DevEnvironment/DevEnvironment.psm1
@@ -13,31 +13,33 @@ function Import-Environment() {
     param ( $file = $(throw "Need a CMD/BAT file to execute"),
             $args = "") 
 
-    $tempFile = [IO.Path]::GetTempFileName()
+	if ([System.IO.File]::Exists($file)) {
+		$tempFile = [IO.Path]::GetTempFileName()
 
-    # Store the output of cmd.exe.  We also ask cmd.exe to output
-    # the environment table after the batch file completes
+		# Store the output of cmd.exe.  We also ask cmd.exe to output
+		# the environment table after the batch file completes
 
-    cmd /c " `"$file`" $args && set > `"$tempFile`" "
+		cmd /c " `"$file`" $args && set > `"$tempFile`" "
 
-    ## Go through the environment variables in the temp file.
-    ## For each of them, set the variable in our local environment.
-    remove-item -path env:*
-    Get-Content $tempFile | Foreach-Object {
-        if($_ -match "^(.*?)=(.*)$") {
-            $n = $matches[1]
-            if ($n -eq "prompt") {
-                # Ignore: Setting the prompt environment variable has no
-                #         connection to the PowerShell prompt
-            } elseif ($n -eq "title") {
-                $host.ui.rawui.windowtitle = $matches[2];
-                set-item -path "env:$n" -value $matches[2];
-            } else {
-                set-item -path "env:$n" -value $matches[2];
-            }
-        }
+		## Go through the environment variables in the temp file.
+		## For each of them, set the variable in our local environment.
+		remove-item -path env:*
+		Get-Content $tempFile | Foreach-Object {
+			if($_ -match "^(.*?)=(.*)$") {
+				$n = $matches[1]
+				if ($n -eq "prompt") {
+					# Ignore: Setting the prompt environment variable has no
+					#         connection to the PowerShell prompt
+				} elseif ($n -eq "title") {
+					$host.ui.rawui.windowtitle = $matches[2];
+					set-item -path "env:$n" -value $matches[2];
+				} else {
+					set-item -path "env:$n" -value $matches[2];
+				}
+			}
+		}
+		Remove-Item $tempFile
     }
-    Remove-Item $tempFile
 }
 
 function Set-DevEnvironment() {
